requirements.txt
pandas==2.2.3
numpy==2.2.4
boto3==1.37.36
botocore==1.37.36
psycopg2-binary==2.9.3
sqlalchemy==1.4.39
PyPDF2==2.10.5
requests==2.28.1
matplotlib==3.10.1
seaborn==0.13.2
